requirements.txt
litellm[proxy]
python-dotenv
